Fixes #1
This fix uses `sublime.set_timeout` to allow the `update_tabs_visible`
method to be asynchronously in Sublime Text 2. Doing this with a 200ms
delay seems to resolve the issue where the views list for the active
window is not being updated when the user has double clicked a file in
the project sidebar.